refactor: move BeanInitializationException to context package
Moved BeanInitializationException from the security to the context package for better organization and clarity. Updated all relevant import statements to reflect this change.
